Refactor import paths for test constants in game tests
- Updated import statements in game.test.ts to reference the new location of test constants, ensuring consistency and improving code organization.
