+++ pvm/ts/src/engine/testConstants.ts
@@ -1,6 +1,6 @@
 import { GameOptions, Positions } from "@bitcoinbrisbane/block52";
 import { ethers } from "ethers";
-import TexasHoldemGame from "../../pvm/ts/src/engine/texasHoldem";
+import TexasHoldemGame from "./texasHoldem";
 
 // Constants for testing
 export const ONE_TOKEN = 100000000000000000n;
